requirements.txt
Django
python-feedparser
